10 [feat] : 홈 페이지 폴더 구조 및 컴포넌트 분리
